refactor: added files for routes for readability
