+++ .env.example
@@ -1,4 +1,5 @@
 IDP_URL="This would be the URL of the Identity Provider, e.g. https://test.ansattporten.no"
+IDP_NAME="idporten"
 CLIENT_ID="Id of your client registered in the Identity Provider"
 CLIENT_SECRET="Secret of your client registered in the Identity Provider"
