+++ requirements.txt
@@ -1,8 +1,11 @@
 annotated-types==0.7.0
 anyio==4.4.0
+Authlib==1.3.1
 certifi==2024.7.4
+cffi==1.16.0
 charset-normalizer==3.3.2
 click==8.1.7
+cryptography==43.0.0
 dnspython==2.6.1
 email_validator==2.2.0
 fastapi==0.111.1
@@ -12,11 +15,14 @@ httpcore==1.0.5
 httptools==0.6.1
 httpx==0.27.0
 idna==3.7
+itsdangerous==2.2.0
 Jinja2==3.1.4
+jwt==1.3.1
 markdown-it-py==3.0.0
 MarkupSafe==2.1.5
 mdurl==0.1.2
 pillow==10.4.0
+pycparser==2.22
 pydantic==2.8.2
 pydantic_core==2.20.1
 Pygments==2.18.0
